Feature/mise en place des tags

+++ src/main/java/com/timeseries/seriestemporelles/controller/SeriesController.java
@@ -13,7 +13,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.Collections;
 import java.util.List;
 
 @RestController
@@ -29,19 +28,33 @@ public class SeriesController {
     @GetMapping("/series")
     private List getAllSeries() { return seriesService.getAllSeries(); }
 
-    @GetMapping("/serie/{id}")
-    private SeriesModel getSerieById(@PathVariable("id") Integer id) {
-        return seriesService.getSerieById(id).orElseThrow(() ->
-            new ResourceNotFoundException("Serie: " + id + " is not found."));
+    @GetMapping("/serie/{id}/user_id={user_id}")
+    private SeriesModel getSerieById(@PathVariable("id") Integer id,
+                                     @PathVariable("user_id") Integer userId) {
+        UserModel user = userService.getUserById(userId).orElseThrow(() ->
+                new ResourceNotFoundException("User: " + userId + "not found."));
+
+        SeriesModel serie = seriesService.getSerieById(id).orElseThrow(() ->
+                new ResourceNotFoundException("Serie: " + id + " is not found."));
+
+        UserSeriesModel userSerie = userSerieService.getUserSerieByUserSerie(user, serie).orElseThrow(() ->
+                new ResourceNotFoundException("UserSerie is not found."));
+
+        if(userSerie.getUserPrivilege() == UserPrivilage.WRITE_PRIVILAGE) {
+            return serie;
+        }
+        return null;
     }
 
     @PostMapping("/serie/{id}")
-    private ResponseEntity createSerie(@PathVariable("id") Integer id, @RequestBody SeriesModel series) {
+    private ResponseEntity createSerie(@PathVariable("id") Integer id,
+                                       @RequestBody SeriesModel series) {
         try {
             series.setLastUpdatedDate();
             seriesService.saveOrUpdate(series);
 
-            UserModel user = userService.getUserById(id).orElseThrow(() ->new ResourceNotFoundException("User: " + id + "is not found."));
+            UserModel user = userService.getUserById(id).orElseThrow(() ->
+                    new ResourceNotFoundException("User: " + id + " is not found."));
 
             UserSeriesModel userSeries = new UserSeriesModel();
             userSeries.setUsers(user);
@@ -56,10 +69,50 @@ public class SeriesController {
         return new ResponseEntity("New series created with id: " + series.getId(), HttpStatus.CREATED);
     }
 
-    @DeleteMapping("/serie/{id}")
-    private ResponseEntity deleteById(@PathVariable("id") int id) {
+    @PutMapping("/serie/{id}/user_id={user_id}")
+    private ResponseEntity updateSerie(@PathVariable("id") Integer id,
+                                       @PathVariable("user_id") Integer userId,
+                                       @RequestBody SeriesModel series) {
         try {
-            seriesService.delete(id);
+            UserModel user = userService.getUserById(userId).orElseThrow(() ->
+                    new ResourceNotFoundException("User: " + userId + "not found."));
+
+            SeriesModel serie = seriesService.getSerieById(id).orElseThrow(() ->
+                    new ResourceNotFoundException("Serie: " + id + " is not found."));
+
+            UserSeriesModel userSerie = userSerieService.getUserSerieByUserSerie(user, serie).orElseThrow(() ->
+                    new ResourceNotFoundException("UserSerie is not found."));
+
+            if(userSerie.getOwner()) {
+                seriesService.saveOrUpdate(series);
+            } else {
+                return new ResponseEntity("User doesn't have permission", HttpStatus.BAD_REQUEST);
+            }
+        } catch (IllegalArgumentException exception) {
+            return ResponseEntity.badRequest().build();
+        }
+        return new ResponseEntity("New series created with id: " + series.getId(), HttpStatus.CREATED);
+    }
+
+    @DeleteMapping("/serie/{id}/user_id={user_id}")
+    private ResponseEntity deleteById(@PathVariable("id") Integer id,
+                                      @PathVariable("user_id") Integer userId) {
+        try {
+            UserModel user = userService.getUserById(userId).orElseThrow(() ->
+                    new ResourceNotFoundException("User: " + userId + "not found."));
+
+            SeriesModel serie = seriesService.getSerieById(id).orElseThrow(() ->
+                    new ResourceNotFoundException("Serie: " + id + " is not found."));
+
+            UserSeriesModel userSerie = userSerieService.getUserSerieByUserSerie(user, serie).orElseThrow(() ->
+                    new ResourceNotFoundException("UserSerie is not found."));
+
+            if(userSerie.getOwner()) {
+                seriesService.delete(id);
+                userSerieService.delete(userSerie.getId());
+            } else {
+                return new ResponseEntity("User doesn't have permission", HttpStatus.BAD_REQUEST);
+            }
         } catch (Exception exception) {
             return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
         }
